Port useReactionsReader hook to ReactionsReader class.
